fix: import error by renaming types
ImportError: cannot import name 'GenericAlias' from partially initialized module 'types' (most likely due to a circular import) (/Users/hslee/Documents/mcp_atlassian/src/mcp_atlassian/types.py)
